+++ .formatter.exs
@@ -2,6 +2,7 @@ locals_without_parens = [
   defvocab: 2,
   defnamespace: 2,
   defnamespace: 3,
+  act_as_namespace: 1,
   def_facet_constraint: 2,
   def_applicable_facet: 1,
   bgp: 1,